# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Event Booking Api",
        "description": "This is the API documentation for Event Booking API",
        "contact": {
            "email": "himanshumishra.tmu@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Localhost server"
        }
    ],
    "paths": {
        "/api/attendees": {
            "post": {
                "tags": [
                    "Attendees"
                ],
                "summary": "Register a new attendee",
                "operationId": "caa45abe11a508c27e8de3f2c80cd64a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Attendee registered",
                        "headers": {
                            "Content-Type": {
                                "schema": {
                                    "type": "string",
                                    "example": "application/json"
                                }
                            },
                            "Accept": {
                                "schema": {
                                    "type": "string",
                                    "example": "application/json"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "headers": {
                            "Content-Type": {
                                "schema": {
                                    "type": "string",
                                    "example": "application/json"
                                }
                            },
                            "Accept": {
                                "schema": {
                                    "type": "string",
                                    "example": "application/json"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Validation error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/bookings": {
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Book an event",
                "operationId": "893be3fcfe94b319a05e6e01eed19c8e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "event_id": {
                                        "type": "integer"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Booking created",
                        "headers": {
                            "Location": {
                                "description": "Location of the new booking",
                                "schema": {
                                    "type": "string",
                                    "format": "uri"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "event_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "attendee_id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Event not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Duplicate booking",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Duplicate booking"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "List all events",
                "operationId": "1031372a8175e66ba1b46a820fe040d0",
                "parameters": [
                    {
                        "name": "country",
                        "in": "query",
                        "description": "Filter by country",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of events",
                        "headers": {
                            "Content-Type": {
                                "schema": {
                                    "type": "string",
                                    "example": "application/json"
                                }
                            },
                            "Accept": {
                                "schema": {
                                    "type": "string",
                                    "example": "application/json"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "headers": {
                            "Content-Type": {
                                "schema": {
                                    "type": "string",
                                    "example": "application/json"
                                }
                            },
                            "Accept": {
                                "schema": {
                                    "type": "string",
                                    "example": "application/json"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Events"
                ],
                "summary": "Create a new event",
                "operationId": "ac2effdd73571c2df8cbc83f879d318e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "capacity": {
                                        "type": "integer"
                                    },
                                    "location": {
                                        "type": "string"
                                    },
                                    "country": {
                                        "type": "string"
                                    },
                                    "state": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Event created",
                        "headers": {
                            "Content-Type": {
                                "schema": {
                                    "type": "string",
                                    "example": "application/json"
                                }
                            },
                            "Accept": {
                                "schema": {
                                    "type": "string",
                                    "example": "application/json"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Bad Request",
                        "headers": {
                            "Content-Type": {
                                "schema": {
                                    "type": "string",
                                    "example": "application/json"
                                }
                            },
                            "Accept": {
                                "schema": {
                                    "type": "string",
                                    "example": "application/json"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Validation error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/events/{id}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get event details",
                "operationId": "f6e17a4fe670cb397894b247c6c5d75f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event details",
                        "headers": {
                            "Content-Type": {
                                "schema": {
                                    "type": "string",
                                    "example": "application/json"
                                }
                            },
                            "Accept": {
                                "schema": {
                                    "type": "string",
                                    "example": "application/json"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "headers": {
                            "Content-Type": {
                                "schema": {
                                    "type": "string",
                                    "example": "application/json"
                                }
                            },
                            "Accept": {
                                "schema": {
                                    "type": "string",
                                    "example": "application/json"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Event not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Events"
                ],
                "summary": "Update an event",
                "operationId": "101a7775adde6d5a249038616701d479",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "capacity": {
                                        "type": "integer"
                                    },
                                    "location": {
                                        "type": "string"
                                    },
                                    "country": {
                                        "type": "string"
                                    },
                                    "state": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Event updated",
                        "headers": {
                            "Content-Type": {
                                "schema": {
                                    "type": "string",
                                    "example": "application/json"
                                }
                            },
                            "Accept": {
                                "schema": {
                                    "type": "string",
                                    "example": "application/json"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "headers": {
                            "Content-Type": {
                                "schema": {
                                    "type": "string",
                                    "example": "application/json"
                                }
                            },
                            "Accept": {
                                "schema": {
                                    "type": "string",
                                    "example": "application/json"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Event not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Events"
                ],
                "summary": "Delete an event",
                "operationId": "7921b70f326e3bc012507123523e4ee3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Event deleted",
                        "headers": {
                            "Content-Type": {
                                "schema": {
                                    "type": "string",
                                    "example": "application/json"
                                }
                            },
                            "Accept": {
                                "schema": {
                                    "type": "string",
                                    "example": "application/json"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "headers": {
                            "Content-Type": {
                                "schema": {
                                    "type": "string",
                                    "example": "application/json"
                                }
                            },
                            "Accept": {
                                "schema": {
                                    "type": "string",
                                    "example": "application/json"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Event not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Attendees",
            "description": "Attendees"
        },
        {
            "name": "Bookings",
            "description": "Bookings"
        },
        {
            "name": "Events",
            "description": "Events"
        }
    ]
}